pyproject.toml
[tool.poetry]
name = "socketio_asyncapi"
version = "0.1.0"
description = "SocketIO with type checking and auto-generate Asyncapi documentation"
authors = ["Marcel van de Weert <mvdweert@bitlab.nl>" ]
license = "MIT"
readme = "README.md"
# packages = [{include = "src/socketio_asyncapi"}]

[tool.poetry.dependencies]
python = "^3.8.10"
python-socketio = "^5.3.2"
pydantic = {extras = ["email"], version = "^1.10.4"}
pyyaml = "^6.0"
python-engineio = "^4.3.4"
loguru = ">=0.6,<0.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-asyncio =  "^0.23.3"
pytest-cov = "^4.0.0"
autopep8 = "^2.0.0"
codespell = "^2.2.2"
redis = ">=4.4,<6.0"
pylint = ">=2.15.8,<4.0.0"
hypothesis-auto = "^1.1.4"
uvicorn = "^0.25.0"
aiohttp = "^3.9.1"
pytest-aiohttp = "^1.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
